refactor: folders schema
