+++ REGISTER_MS/.gitignore
@@ -1,3 +1,4 @@
 node_modules
+databaseMock
 coverage/
 .env
