+++ REGISTER_MS/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "template",
+  "name": "register-ms",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
