+++ REGISTER_MS/src/adapters/exceptions/index.ts
@@ -1,0 +1,1 @@
+export * from './Exceptions'
