+++ REGISTER_MS/src/adapters/repository/index.ts
@@ -1,0 +1,1 @@
+export * from './UserRepository'
