+++ REGISTER_MS/src/adapters/services/index.ts
@@ -1,0 +1,2 @@
+export * from './ServiceCreateUser'
+export * from './ServiceDeleteAllUsers'
